+++ src/main/java/ru/sbt/mipt/oop/room/Door.java
@@ -1,4 +1,4 @@
-package ru.sbt.mipt.oop;
+package ru.sbt.mipt.oop.room;
 
 public class Door {
     private final String id;
